requirements.txt
dill==0.3.2
jinja2==2.11.2
Keras==2.1.6
matplotlib==3.0.3
numpy==1.18.5
scipy==1.4.1
six==1.12.0
tensorboard==2.2.0
tensorflow_gpu==2.2.0
torch==1.2.0
zmq==0.0.0
gpustat
